Bump package version to 3.3.0
The previous commit was wrong, this is not a patch release.

_Technically_ it might even be a breaking change with regard to the
types, because due to svelte the input type and output types are the
same, so the output type is now less specific and can now be null.

In practice, the component will still not actually return null.
Hopefully this isn't problematic, though this doesn't feel like it
warrants a major version bump.